Refactor CustomGoogleMapWidget to fix camera animation on first call
Previously, the camera animation on the first call was not being executed due to a commented-out line. This commit re-enables the `isFirstCall` flag to ensure that the camera animates correctly on the first location update. This change improves the user experience by ensuring the map view updates smoothly when the location is first obtained.